backend/src/requirements/prod.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements/prod.txt requirements/prod.in
#
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
boto3==1.40.1
    # via -r requirements/base.in
botocore==1.40.1
    # via
    #   boto3
    #   s3transfer
certifi==2025.7.14
    # via requests
charset-normalizer==3.4.2
    # via requests
django==5.2.4
    # via
    #   -r requirements/base.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.7.0
    # via -r requirements/base.in
djangorestframework==3.16.0
    # via -r requirements/base.in
factory-boy==3.3.3
    # via -r requirements/base.in
faker==37.4.2
    # via factory-boy
gunicorn==23.0.0
    # via -r requirements/prod.in
idna==3.10
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
packaging==25.0
    # via gunicorn
psycopg2-binary==2.9.10
    # via -r requirements/base.in
python-dateutil==2.9.0.post0
    # via botocore
requests==2.32.4
    # via -r requirements/base.in
s3transfer==0.13.1
    # via boto3
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via faker
urllib3==2.5.0
    # via
    #   botocore
    #   requests
